+++ pages/api/spreadsheet/urls.ts
@@ -1,12 +1,13 @@
 import { google } from "googleapis";
 import type { NextApiRequest, NextApiResponse } from "next";
+
 import {
   GCP_CLIENT_EMAIL,
   GCP_CLIENT_ID,
   GCP_PRIVATE_KEY,
   GCP_WRITE_SPREADSHEET_SCOPES,
   URLS_SPREADSHEET_ID,
-} from "../../src/server/constants";
+} from "@server/constants";
 
 export default async function handler(
   req: NextApiRequest,
@@ -41,7 +42,33 @@ export default async function handler(
       shortUrl,
       longUrl,
     });
-  } else {
-    res.send("Not Found");
+  } else if (req.method === "GET") {
+    const auth = new google.auth.GoogleAuth({
+      credentials: {
+        client_email: GCP_CLIENT_EMAIL,
+        client_id: GCP_CLIENT_ID,
+        private_key: GCP_PRIVATE_KEY,
+      },
+      scopes: GCP_WRITE_SPREADSHEET_SCOPES,
+    });
+    const sheets = google.sheets({
+      auth,
+      version: "v4",
+    });
+    const response = await sheets.spreadsheets.values.batchGet({
+      spreadsheetId: URLS_SPREADSHEET_ID,
+      ranges: ["Sheet1!A2:C"],
+    });
+    const data = response.data;
+    const values = data.valueRanges?.[0].values;
+
+    const urls = (values || []).map((_, index) => {
+      return {
+        shortUrl: values?.[index]?.[0],
+        longUrl: values?.[index]?.[1],
+      };
+    });
+
+    res.status(201).json(urls.filter(Boolean));
   }
 }
